EEG_Avatar_Feedback_Loop/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
brainflow>=5.0.0
tensorflow>=2.8.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pillow>=8.3.0
mediapipe>=0.8.9
pandas>=1.3.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
moviepy>=1.0.3 